feat(ai-models): extract @grida/ai-models package
Carves the AI model catalogue (text-model specs + tier vocabulary, image,
audio, and image-tool cards) out of editor/lib/ai/ into a standalone
workspace package. Editor side keeps the gateway + BYOK provider seam and
re-exports the catalogue under the original names so call sites are
unchanged.
