+++ packages/grida-ai-models/README.md
@@ -1,0 +1,128 @@
+# @grida/ai-models
+
+A standalone model catalog.
+
+This package publishes typed data for AI model selection, display, pricing, and
+size validation. It does not create provider clients, make network requests,
+enforce billing, or decide access. Its scope ends at exported objects, types,
+and lookup helpers.
+
+## Related
+
+- [Model pricing docs](https://grida.co/docs/models/pricing)
+- [AI models catalog](https://grida.co/ai/models)
+
+## Contents
+
+- Agentic model: `nano`, `mini`, `pro`, and `max`
+- Text model specs: labels, modality, context windows, output limits, and token
+  pricing
+- Image generation model cards: labels, vendors, speed hints, supported sizes,
+  size constraints, defaults, and pricing
+- Audio generation model cards
+- Image tool model cards, such as background removal and upscaling
+- Shared discriminator types for providers, vendors, speed labels, and pricing
+  schemes
+
+## Usage
+
+The package's primary export is the `models` namespace, available as either
+a default or a named import. `TIER_MODEL_IDS` plus the `ModelTier` /
+`TierModelId` types live at the top level alongside it.
+
+```ts
+import models, { TIER_MODEL_IDS } from "@grida/ai-models";
+
+const proModelId = TIER_MODEL_IDS.pro;
+const proModel = models.text.byTier.pro;
+const spec = models.text.modelSpecById("claude-sonnet-4.6");
+
+const imageModel = models.image.models["openai/gpt-image-2"];
+const compactImageModel = imageModel && models.image.toCompact(imageModel);
+
+const audioModel = models.audio.models["google/lyria-3"];
+const upscaleTool = models.image_tools.models["nightmareai/real-esrgan"];
+```
+
+## Text Models
+
+Text models live under `models.text` and are split into two tables:
+
+- `TIER_MODEL_IDS`: maps each `ModelTier` to a model id (top-level export
+  from `./tiers`)
+- `models.text.catalog`: maps model ids to `models.text.ModelSpec` objects
+
+`models.text.byTier` is derived from those two tables and exposes the
+resolved `Record<ModelTier, ModelSpec>`.
+
+Each `ModelSpec` contains:
+
+- `id`
+- `label`
+- `multimodal`
+- `contextWindow`
+- `outputLimit`
+- `cost`
+
+Token costs are stored as USD per 1 million tokens.
+
+## Media Models
+
+Media model data lives under the `models` namespace:
+
+- `models.image`
+- `models.audio`
+- `models.image_tools`
+
+Image cards can describe both preset sizes and continuous size constraints.
+When both are present, `constraints` is the validation envelope and `sizes` is a
+set of suggested presets or pricing anchors.
+
+Image pricing is a discriminated union:
+
+- `per_image_tiered`: quality and size based image prices
+- `per_image_flat`: one price per image
+- `per_token`: token rates for input and output
+
+Audio models currently use flat per-run pricing. Image tools use flat
+per-invocation pricing.
+
+## Lookups
+
+`models.text.modelSpecById(modelId)` accepts:
+
+- Full ids, such as `openai/gpt-5.4-mini`
+- Bare ids, such as `gpt-5.4-mini`
+- Date-suffixed provider ids, such as `gpt-5.4-mini-2025-08-07`
+
+`models.image.findImageModelCard(model)` accepts:
+
+- Full image model ids
+- Bare image model ids when the match is unambiguous
+- The deprecated `{ provider, modelId }` wrapper shape
+
+## Updating The Catalog
+
+To add or update a text model (or any image / audio / image-tool model),
+edit `src/models.ts`. That file is the central catalogue and also the
+type source — `models.text.CatalogId` is derived from the text-model table.
+
+To change a tier mapping, update `TIER_MODEL_IDS` in `src/tiers.ts`. The
+mapped id must already exist in the text catalogue; the compiler enforces
+this because `TIER_MODEL_IDS` is typed against `models.text.CatalogId`.
+
+Keep the stored data literal and portable:
+
+- Use model ids as stable keys.
+- Store real published prices, not application-specific estimates, except for
+  `avg_cost_usd`, which is explicitly a coarse invocation estimate.
+- Keep provider and vendor values as data labels. This package should not
+  import SDKs or contain routing logic.
+- Prefer adding explicit types before widening existing ones.
+
+## Scripts
+
+```sh
+pnpm build
+pnpm typecheck
+```
