+++ packages/grida-ai-models/src/index.ts
@@ -1,0 +1,10 @@
+/**
+ * `@grida/ai-models` — typed AI model catalogue. See README for the
+ * package contract.
+ *
+ * @module
+ */
+
+export * from "./tiers";
+export * from "./models";
+export { default } from "./models";
